requirements.txt
Flask==0.10.1
Flask-Admin==1.4.0
Flask-Principal==0.4.0
Flask-SQLAlchemy==2.1
Flask-Security==1.7.5
